Move schema_validated and _readable_offset to documents from _misc
