+++ datacube/utils/__init__.py
@@ -20,6 +20,8 @@ from .documents import (
     get_doc_offset_safe,
     netcdf_extract_string,
     without_lineage_sources,
+    schema_validated,
+    _readable_offset,
 )
 from .math import (
     unsqueeze_data_array,
@@ -29,9 +31,7 @@ from .math import (
 )
 from ._misc import (
     DatacubeException,
-    schema_validated,
     gen_password,
-    _readable_offset,
 )
